# [improve][broker] Replaced checkBackloggedCursors with checkBackloggedCursor(single subscription check) upon subscription