@mattisonchao 2023-01-28
Question:
When this logic is removed, if the cursor has no consumers and it is set to an `inactive` state by `updateStats`. Will a new subscription activate the cursor?