@heesung-sn 2023-01-28
Also,

If we still want to call this operation, checkBackloggedCursors asynchronously but more frequently, we could run such async operations in a separate thread and run it more frequently, every x milliseconds(out side of updateStats), if there are any requests (with deduplication).

// Example 
'''
AsyncOpManager:

var dedupedOps<Op>

void schedule(Op op) // dedup/add op to dedupedOps.

void start(){

 while(true)
  // wakeup every x milliseconds s(configurable) and run the deduped ops

}
'''