@heesung-sn 2023-01-29
> I think it would be better to call it only for the single subscription, instead of for all of them
Updated the code to check the single subscription.