@mattisonchao 2023-01-29
@heesung-sn  Please change the PR title.

>Unknown scope "pulsar" found in pull request title "[improve][pulsar] Replaced checkBackloggedCursors with checkBackloggedCursor(single subscription check) from subscribe". Use one of the available scopes: admin, bk, broker, build, ci, cli, client, doc, fn, io, meta, misc, ml, monitor, offload, proxy, schema, sec, site, sql, storage, test, txn, ws, zk.